+++ test/ColdResumptionTest.cpp
@@ -81,7 +81,7 @@ class HelloResumeHandler : public ColdResumeHandler {
 
   Reference<Subscriber<Payload>> handleRequesterResumeStream(
       std::string streamToken,
-      uint32_t consumerAllowance) override {
+      size_t consumerAllowance) override {
     CHECK(subscribers_.find(streamToken) != subscribers_.end());
     VLOG(1) << "Resuming " << streamToken << " stream with allowance "
             << consumerAllowance;
